requirements.txt
pandas
numpy
yfinance
python-binance
dukascopy-python
pyyaml
pyarrow
fastparquet
requests
